+++ src/common/mutable.rs
@@ -29,7 +29,7 @@ pub struct MutableItem {
 
 impl MutableItem {
     /// Create a new mutable item from a signing key, value, sequence number and optional salt.
-    pub fn new(signer: SigningKey, value: &[u8], seq: i64, salt: Option<&[u8]>) -> Self {
+    pub fn new(signer: &SigningKey, value: &[u8], seq: i64, salt: Option<&[u8]>) -> Self {
         let signable = encode_signable(seq, value, salt);
         let signature = signer.sign(&signable);